add simple perf script
The goal of this is not to measure in detail, but get a ballpark against
the current errol implementation to make sure we are roughly in line.

I don't expect amazing performance just yet since the implementation is
the generic 128-bit backend and there is little optimization in the
character emitting.